Prod-0: Rename model folder to domian
