+++ src/main/java/com/indiraactive/fulfillmentplatform/domain/RunHistoryModel.java
@@ -1,4 +1,4 @@
-package com.indiraactive.fulfillmentplatform.model;
+package com.indiraactive.fulfillmentplatform.domain;
 
 import java.util.Date;
