+++ src/main/java/com/indiraactive/fulfillmentplatform/domain/ScheduledTaskHistoryModel.java
@@ -1,4 +1,4 @@
-package com.indiraactive.fulfillmentplatform.model;
+package com.indiraactive.fulfillmentplatform.domain;
 
 import java.util.Date;
 import java.util.List;
