+++ src/main/java/com/indiraactive/fulfillmentplatform/viewModel/ScheduleHistoryViewModel.java
@@ -1,7 +1,7 @@
 package com.indiraactive.fulfillmentplatform.viewModel;
 
 import com.indiraactive.fulfillmentplatform.dal.supplier.SupplierRepository;
-import com.indiraactive.fulfillmentplatform.model.ScheduledTaskHistoryModel;
+import com.indiraactive.fulfillmentplatform.domain.ScheduledTaskHistoryModel;
 import com.indiraactive.fulfillmentplatform.dal.scheduledTask.ScheduledTaskJpa;
 
 import java.util.LinkedList;
